# swagger/swagger.yml
openapi: 3.0.0
paths:
...
  /plantilla/{id}:
    get:
      operationId: PlantillaController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_1
    put:
      operationId: PlantillaController_put
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlantillaDto'
      responses:
        '200':
          description: ''
      tags: *ref_1
    delete:
      operationId: PlantillaController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_1
...
  /tipo_contenido:
    post:
      operationId: TipoContenidoController_post
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TipoContenidoDto'
      responses:
        '201':
          description: ''
      tags: &ref_3
        - tipo_contenido
    get:
      operationId: TipoContenidoController_getAll
      parameters:
        - name: query
          required: false
          in: query
          description: query - Filter. e.g. col1:v1,col2:v2 …
          schema:
            type: string
        - name: fields
          required: false
          in: query
          description: fields - Fields returned. e.g. col1,col2 …
          schema:
            type: string
        - name: sortby
          required: false
          in: query
          description: sortby - Sorted-by fields. e.g. col1,col2 …
          schema:
            type: string
        - name: order
          required: false
          in: query
          description: >-
            order - Order corresponding to each sortby field, if single value,
            apply to all sortby fields. e.g. desc,asc …
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: limit - Limit the size of result set. Must be an integer
          schema:
            type: string
        - name: offset
          required: false
          in: query
          description: offset - Start position of result set. Must be an integer
          schema:
            type: string
        - name: populate
          required: false
          in: query
          description: populate - show subqueries. Must be a boolean
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_3
  /tipo_contenido/{id}:
    get:
      operationId: TipoContenidoController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_3
    put:
      operationId: TipoContenidoController_put
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TipoContenidoDto'
      responses:
        '200':
          description: ''
      tags: *ref_3
    delete:
      operationId: TipoContenidoController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_3
  /tipo_plantilla:
    post:
      operationId: TipoPlantillaController_post
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TipoPlantillaDto'
      responses:
        '201':
          description: ''
      tags: &ref_4
        - tipo_plantilla
    get:
      operationId: TipoPlantillaController_getAll
      parameters:
        - name: query
          required: false
          in: query
          description: query - Filter. e.g. col1:v1,col2:v2 …
          schema:
            type: string
        - name: fields
          required: false
          in: query
          description: fields - Fields returned. e.g. col1,col2 …
          schema:
            type: string
        - name: sortby
          required: false
          in: query
          description: sortby - Sorted-by fields. e.g. col1,col2 …
          schema:
            type: string
        - name: order
          required: false
          in: query
          description: >-
            order - Order corresponding to each sortby field, if single value,
            apply to all sortby fields. e.g. desc,asc …
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: limit - Limit the size of result set. Must be an integer
          schema:
            type: string
        - name: offset
          required: false
          in: query
          description: offset - Start position of result set. Must be an integer
          schema:
            type: string
        - name: populate
          required: false
          in: query
          description: populate - show subqueries. Must be a boolean
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_4
  /tipo_plantilla/{id}:
    get:
      operationId: TipoPlantillaController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_4
    put:
      operationId: TipoPlantillaController_put
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TipoPlantillaDto'
      responses:
        '200':
          description: ''
      tags: *ref_4
    delete:
      operationId: TipoPlantillaController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_4
info:
  title: plantillas_crud
  description: API CRUD sistema de plantillas
  version: '1.0'
  contact: {}
...
servers: []
components:
  schemas:
...
    PlantillaDto:
      type: object
      properties:
        tipo_plantilla_id:
          type: string
        secciones:
          type: array
          items:
            type: string
        sistema_id:
          type: string
        grupo_id:
          type: string
        version:
          type: string
        nombre:
          type: string
        uid:
          type: boolean
        estado:
          type: boolean
        activo:
          type: boolean
        fecha_creacion:
          format: date-time
          type: string
        fecha_modificacion:
          format: date-time
          type: string
      required:
        - tipo_plantilla_id
        - secciones
        - sistema_id
        - grupo_id
        - version
        - nombre
        - uid
        - estado
        - activo
        - fecha_creacion
        - fecha_modificacion
...
    TipoContenidoDto:
      type: object
      properties:
        nombre:
          type: string
        descripcion:
          type: string
        codigo_abreviacion:
          type: string
      required:
        - nombre
        - descripcion
        - codigo_abreviacion
    TipoPlantillaDto:
      type: object
      properties:
        nombre:
          type: string
        descripcion:
          type: string
        codigo_abreviacion:
          type: string
      required:
        - nombre
        - descripcion
        - codigo_abreviacion